assert check rewrite.
git-svn-id: http://southwind:2481/svn/cpp_lib@116 287b3242-7fab-264f-8401-8509467ab285

+++ WinThread/WinBarrier.hpp
@@ -64,7 +64,8 @@ public:
 		}
 		else
 		{
-			assert(waitedThreads.size() == (maxCount-1));
+			assert(waitedThreads.size() == 
+				static_cast<size_t>(maxCount-1));
 
 			count = maxCount;
 			for (std::vector<HANDLE>::iterator itor = waitedThreads.begin();
